Greenify pre-commit by fixing the following pylint checks:
pylint with optional checks..............................................Passed
- hook id: pylint
- duration: 61.46s

************* Module l10n_it_fatturapa
l10n_it_fatturapa/views/partner_view.xml:51: [W7940(dangerous-view-replace-wo-priority), ]  Dangerous use of "replace" from view with priority 0 < 99
************* Module l10n_it_abicab
l10n_it_abicab/views/abicab_view.xml:29: [W7940(dangerous-view-replace-wo-priority), ]  Dangerous use of "replace" from view with priority 0 < 99
************* Module l10n_it_intrastat_statement.models.intrastat_statement_purchase_section
l10n_it_intrastat_statement/models/intrastat_statement_purchase_section.py:58: [R7980(consider-merging-classes-inherited), IntrastatStatementPurchaseSection1] Consider merging classes inherited to "account.intrastat.statement.purchase.section" from intrastat_statement_purchase_section.py:212, intrastat_statement_purchase_section.py:357, intrastat_statement_purchase_section.py:428.
************* Module l10n_it_intrastat_statement.models.intrastat_statement_sale_section
l10n_it_intrastat_statement/models/intrastat_statement_sale_section.py:26: [R7980(consider-merging-classes-inherited), IntrastatStatementSaleSection1] Consider merging classes inherited to "account.intrastat.statement.sale.section" from intrastat_statement_sale_section.py:172, intrastat_statement_sale_section.py:300, intrastat_statement_sale_section.py:371.

pylint with mandatory checks.............................................Failed
- hook id: pylint
- exit code: 4

************* Module l10n_it_fatturapa
l10n_it_fatturapa/views/partner_view.xml:51: [W7940(dangerous-view-replace-wo-priority), ]  Dangerous use of "replace" from view with priority 0 < 99
************* Module l10n_it_abicab
l10n_it_abicab/views/abicab_view.xml:29: [W7940(dangerous-view-replace-wo-priority), ]  Dangerous use of "replace" from view with priority 0 < 99
